Add option to filter objects
Some orgs have lots of objects and it can be annoying to find the correct object for query or load. to address this, we have added a quick filter to reduce the list of objects to look through.

resolves #1132
